# pgrest/resources/openapi_v3.yml
openapi: 3.0.1
info:
  title: Tapis PgREST API
  description: The Tapis PgREST API provides a RESTful interface to a managed SQL-db-as-a-service.
  contact:
    name: CICSupport
    email: cicsupport@tacc.utexas.edu
  license:
    name: 3-Clause BSD License
    url: 'https://opensource.org/licenses/BSD-3-Clause'
  version: '1.0.0'
...
servers:
- url: http://localhost:5000
  description: Local Development
- url: https://dev.develop.tapis.io
  description: Develop Instance
#security:
#  - TapisJWT: []
...
paths:
...
  /v3/pgrest/data/{root_url}:
    get:
      tags:
        - Tables
      summary: get_table_rows
      description: List objects in the table with specified root_url {root_url}.
      operationId: get_table_rows
      parameters:
      - name: root_url
        in: path
        description: The root_url parameter of the table.
        required: true
        schema:
          type: string
      - name: limit
        in: query
        description: limit the number of records returned.
        schema:
          type: integer
      - name: offset
        in: query
        description: index (offset) to start list.
        schema:
          type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BasicResponse'
                properties:
                  result:
                    $ref: '#/components/schemas/TableRows'
    post:
      tags:
        - Tables
      summary: add_table_row
      description: Create a new object in the table with specified root_url {root_url}.
      operationId: add_table_row
      parameters:
      - name: root_url
        in: path
        description: The root_url parameter of the table.
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewTableRow'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BasicResponse'
                properties:
                  result:
                    $ref: '#/components/schemas/TableRow'
    put:
      tags:
        - Tables
      summary: update_table_rows
      description: Update multiple rows in a table atomically based on filter, if no filter, update all.
      operationId: update_table_rows
      parameters:
      - name: root_url
        in: path
        description: The root_url parameter of the table.
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMultipleTableRows'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BasicResponse'
  /v3/pgrest/data/{root_url}/{item}:
    get:
      tags:
        - Tables
      summary: get_table_row
      description: Get details about the specific object with id {item} on the table with root_url {root_url}.
      operationId: get_table_row
      parameters:
      - name: root_url
        in: path
        description: The root_url parameter of the table.
        required: true
        schema:
          type: string
      - name: item
        in: path
        description: The id of an item (i.e., a row) on the table whose root_url is given by {root_url}.
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BasicResponse'
                properties:
                  result:
                    $ref: '#/components/schemas/TableRow'
    put:
      tags:
        - Tables
      summary: update_table_row
      description: Update a specific object with id {item} from the table with root_url {root_url}.
      operationId: update_table_row
      parameters:
      - name: root_url
        in: path
        description: The root_url parameter of the table.
        required: true
        schema:
          type: string
      - name: item
        in: path
        description: The id of an item (i.e., a row) on the table whose root_url is given by {root_url}.
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTableRow'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BasicResponse'
                properties:
                  result:
                    $ref: '#/components/schemas/TableRow'
    delete:
      tags:
        - Tables
      summary: delete_table_row
      description: Delete a specific object with id {item} from the table with root_url {root_url}.
      operationId: delete_table_row
      parameters:
      - name: root_url
        in: path
        description: The root_url parameter of the table.
        required: true
        schema:
          type: string
      - name: item
        in: path
        description: The id of an item (i.e., a row) on the table whose root_url is given by {root_url}.
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BasicResponse'


  #=== VIEWS ===#
...
  /v3/pgrest/manage/views/{view_name}:
    get:
      tags:
        - Manage Views
      summary: get_manage_view
      description: Get details of a specific view.
      operationId: get_manage_view
      x-swagger-router-controller: connexContr
      parameters:
      - name: view_name
        in: path
        description: The id of the view.
        required: true
        schema:
          type: string
      - name: details
        in: query
        description: Get additional details about the view
        schema:
          type: boolean
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BasicResponse'
                properties:
                  result:
                    $ref: '#/components/schemas/View'
    delete:
      tags:
        - Manage Views
      summary: delete_view
      description: Delete a specific view. WARNING -- this action cannot be undone.
      operationId: delete_view
      x-swagger-router-controller: connexContr
      parameters:
      - name: view_name
        in: path
        description: The name of the view to delete.
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BasicResponse'
  /v3/pgrest/views/{view_name}:
    get:
      tags:
        - Views
      summary: get_view
      description: List objects in the view with root_url {view_name}.
      operationId: get_view
      parameters:
      - name: view_name
        in: path
        description: The root_url parameter of the view.
        required: true
        schema:
          type: string
      - name: limit
        in: query
        description: limit the number of records returned.
        schema:
          type: integer
      - name: offset
        in: query
        description: index (offset) to start list.
        schema:
          type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BasicResponse'
                properties:
                  result:
                    $ref: '#/components/schemas/TableRows'


  #=== ROLES ===#
...
components:
  schemas:
    BasicResponse:
      type: object
      properties:
        version:
          type: string
          description: Version of the API
        message:
          type: string
          description: Brief description of the response
        status:
          type: string
          description: Whether the request was a success or failure.
          enum: [success, error]


    #=== TABLES ===#
...
    TableRow:
      type: object
    TableRows:
      type: array
      items:
        $ref: '#/components/schemas/TableRow'
    NewTableRow:
      type: object
      properties:
        data:
          type: object
          description: The values of the row to add.
    UpdateTableRow:
      type: object
      description: A JSON object of key/value pairs describing the updates to be made to the item. Each key should be the name of a column and each corresponding value should be the new value to update the column to.
    UpdateMultipleTableRows:
      type: object
      description: Update multiple rows in the same way atomically, where the set of rows to be updated is described with a "where" clause and the update to be performed is defined by a JSON object of column names and column values (the new values to apply to all records).
      properties:
        where:
          type: object
          description: A JSON object describing a where clause of records to update; each record on the table matching the where clause will be updates.
          properties:
            sub_value:
              type: object
              properties:
                operator:
                  type: string
                  description: The operator to use in the where clause when comparing rows to be updated.
                value:
                  description: The value to use in the where clause when comparing rows to be updated.
                  oneOf:
                    - type: integer
                    - type: string
                    - type: boolean
      additionalProperties:
        type: object
      required: [where]


    #=== VIEWS ===#
    View:
      type: object
      properties:
        view_name:
          type: string
          description: The name of the view.
        root_url:
          type: string
          description: The url path used for the root root_url associated with the view.
        endpoints:
          type: string
          description: Which endpoints are available for the view. (This doesn't work/doesn't mean anything)
        manage_view_id:
          type: string
          description: View ID to manage the view with. Serial IDs allocated per tenant.
        tenant_id:
          type: string
          description: The name of the tenant this view is created in.
        comments:
          type: string
          description: Text area to describe view. Returned when calling manage endpoints.
...